+++ src/config.h
@@ -33,12 +33,12 @@
 #pragma mark once
 
 #define PACKAGE	"Mosaic"
-#define	VERSION	"0.1.6"
+#define	VERSION	"0.1.7"
 #define DESCRIPTION "OF Visual Patching Developer Platform"
 #define MOSAIC_WWW "https://mosaic.d3cod3.org/"
 #define TAGS "mosaic,openframeworks,macos,linux,windows,creative-coding,video,audio,graphics,live-coding,visual-programming,creative-coding,transmedia"
 
-#define WINDOW_TITLE "Mosaic 0.1.6"
+#define WINDOW_TITLE "Mosaic 0.1.7"
 #define WINDOW_START_WIDTH 1280
 #define WINDOW_START_HEIGHT 720