+++ src/includes.h
@@ -33,3 +33,9 @@
 #pragma mark once
 
 #include "ofxVisualProgramming.h"
+#include "ofxImGui.h"
+#include "ofxScreenLoggerChannel.h"
+#include "ofxModal.h"
+#include "ofxSimpleHttp.h"
+
+#include "MosaicTheme.h"